Synchronize SSH keys
Bobber relies on SSH keys that are baked into the images to enable
multi-node communication. This forces users to build the image on one
machine, save the image locally, copy it to all remote nodes, and load
the copied image on those hosts. This process is long and tedious, but
by replacing it with a synchronization method, makes it possible to
run the build on each host and not need to copy images remotely.

Signed-Off-By: Robert Clark <roclark@nvidia.com>