Use CUDA image from NGC as base
Bobber currently uses a TensorFlow image from NGC as the base to use some
of the TensorFlow functionality in the tests. While this is efficient,
the Bobber image is quite large (12GB+ at present). By moving to a CUDA
base image and installing TensorFlow inside the Bobber image, it might be
possible to reduce the overall image size by several gigabytes with no
change in functionality or performance. This will require a thorough
investigation of the potential impacts of such a change.

Signed-Off-By: Robert Clark <roclark@nvidia.com>