Add documentation on building from source
To make it easier to test local updates, steps should be added
demonstrating how to run code directly from source, including building a
new wheel if desired.

A simple bash script has also been added to bump the version to a
dev-version including a timestamp, plus automatically generate a wheel
and save it to the dist/ directory.

Signed-Off-By: Robert Clark <roclark@nvidia.com>