Add support for Jenkins
GitHub Actions is limited as a Continuous Integration tool given it
doesn't support GPUs and isn't high-performant. Additionally, using a
GitHub Actions runner on a self-hosted machine for a public repository
exposes several security risks that should be avoided.

Running Jenkins in a secure lab connected to dedicated runners which are
isolated from the external web minimizes many of the risks associated
with GitHub Actions.

Signed-Off-By: Robert Clark <roclark@nvidia.com>