Use multi-stage builds to reduce image size
The Bobber image requires compiling multiple binaries across many
repositories, creating a lot of unnecessary files which bloats the
image. By using multi-stage builds, much of the compilation and
dependency installation can be done in a beefier base image and only the
necessary components can be copied to the final runtime image. This
results in a reduction of several gigabytes in the final image for no
loss of performance or functionality.

Signed-Off-By: Robert Clark <roclark@nvidia.com>