Fix metadata parsing with warnings in logs
Some corner cases exist where a warning is raised in the mdtest logs but
the tests are still able to complete. This throws the parser off as it
would attempt to pull the warning message in with the operation summary
and throw an error that it can't pull values from the warning message.

Signed-Off-By: Robert Clark <roclark@nvidia.com>